third_party/blink/web_tests/external/wpt/html/canvas/element/manual/imagebitmap/canvas-createImageBitmap-resize-expected.txt

This is a testharness.js-based test.
[FAIL] createImageBitmap from a HTMLImageElement of svg with no specified size with resize option.
  assert_array_equals: This pixel should be red. expected property 0 to be 255 but got 0 (expected array [255, 0, 0, 255] got object "0,0,0,0")
Harness: the test ran to completion.
